Imported Upstream version 0.9.4
0.9.4                   - fix parsing of square brackets in string OID labels
evvolve                 - close the file descriptor in parse_config_file
ctrl                    - re-work of decoding infrastructure for CableHome

0.9.3                   - update CMTS MIC calculation per DOCSIS 2.0 (added
evvolve                   SubMgmt settings to CMTS MIC calculation)
ctrl                    - fix decoding of long Strings in SNMP Objects
ctrl                    - add support for long Variable Bindings (>127)
ctrl                    - add support for TLV64 encode/decode for long SNMP
ctrl                      Objects.
evvolve                 - add support for Downstream Channel List TLVs
evvolve                 - add support for TftpModemAddress and TftpTimeout TLVs
ctrl                    - make GenericTLVs easier to decode/encode when
ctrl                      printable strings are used
ctrl                    - fixed memory leaks (for the sake of it)

+++ src/ethermac.c
@@ -48,7 +48,7 @@ p = strchr ( ptr, (int) ':' );
  	}
 
 	memset(fragptr,0,3); 
-	memcpy ( fragptr, ptr, p-ptr );
+	memcpy ( fragptr, ptr, (size_t)(p-ptr) );
 
 	if ( (rval= sscanf ( fragptr, "%x", &fragval)) == 0 ) { 
 		printf("\nInvalid MAC Address %s\n", macstr );
